@@ -1,6 +1,6 @@
 module silence
 
-go 1.23.3
+go 1.24
 
 require (
 	github.com/google/uuid v1.6.0
